limit the user comment delete ability to the user

+++ src/components/CommentsCard/CommentsCard.jsx
@@ -1,17 +1,25 @@
 import "./CommentsCard.css";
 import * as commentsAPI from "../../utilities/comments-api";
 import { useParams } from "react-router-dom";
+//import { useNavigate } from "react-router-dom";
 
-export default function CommentsCard({ comment }) {
-  // TODO:  need to find which user created the comment
+export default function CommentsCard({ comment, user }) {
   const { postId } = useParams();
+  //Use the navigate function to change routes programmatically
+  // const navigate = useNavigate();
 
   async function handleDelete() {
-    alert("Are you sure to delete the comment?");
-    //comment._id
-    const deleteCom = await commentsAPI.deleteComment(comment._id);
-    console.log(deleteCom);
-    // need to refresh it
+    // if the user of the comment is same as the login user
+    if (comment.user === user._id) {
+      alert("Are you sure to delete the comment?");
+      //comment._id
+      const deleteCom = await commentsAPI.deleteComment(comment._id);
+      console.log(deleteCom);
+      // need to refresh it
+      //navigate(`/${postId}`); //not working
+    } else {
+      alert("you don't ...");
+    }
     window.location.href = `/${postId}`;
   }
 
@@ -28,8 +36,7 @@ export default function CommentsCard({ comment }) {
       </p>
       <button onClick={handleDelete}>Delete</button>
       {/* ???????????????????????????????????? */}
-      {/* <p>Author: {comment.user.name}</p>
-      <p>Author Info: {comment.user.email}</p> */}
+      {/* <p>Author: {comment.user.name}</p>*/}
     </div>
   );
 }
